uninstall : added unit tests
